producer/requirements.txt
pandas
faker
kafka-python